src/django/requirements.txt
astroid>=2.4.2
beautifulsoup4>=4.8.1

Django==3.2.12
django-filter>=2.4.0
django-livereload-server>=0.3.2
django-webpack-loader>=0.6.0
django-registration>=3.1.1
djangorestframework>=3.12.1

isort>=5.6.4
jedi>=0.17.2
lazy-object-proxy>=1.4.3
mccabe>=0.6.1
names>=0.3.0
parso>=0.7.1
psycopg2>=2.8.6
pylint>=2.6.0
pytz>=2019.3
six>=1.12.0
soupsieve>=1.9.4
sqlparse>=0.3.0
toml>=0.10.1
tornado>=6.0.3
typed-ast>=1.4.1
wrapt>=1.12.1
whitenoise>=5.2.0

pydantic>=1.7.1
email-validator>=1.1.1

gunicorn>=20.0.4
bs4>=0.0.1

Markdown>=3.3.3
django-annoying>=0.10.6
django-sendgrid-v5>=0.9.0
sendgrid>=6.4.7
premailer>=3.7.0
coverage>=5.3
geojson>=2.5.0
requests>=2.24.0
toolz>=0.11.1

python-dateutil>=2.8.1
environs>=9.3.5

cc-backend-lib>=0.11.0
geomock>=2.0.1
Faker==13.2.0
django-extensions==3.1.5
